### 2019-12-01 #1923 @whedon
[PRE REVIEW]: Annotation Sniffer: A tool to Extract Code Annotations Metrics
**Submitting author:** @phillima (<a href="http://orcid.org/0000-0002-8358-4405">Phyllipe Lima</a>)
**Repository:** <a href="https://github.com/phillima/asniffer/" target ="_blank">https://github.com/phillima/asniffer/</a>
**Version:** v2.0.0
