+++ joss-paper/paper.md
@@ -28,22 +28,22 @@ bibliography: paper.bib
 ---
 
 # Summary
-Enterprise Java frameworks and APIs such as JPA (Java Persistence API), Spring, EJB (Enterprise Java Bean), and JUnit make extensive use of code annotations as means to allow applications to configure custom metadata and execute specific behavior. Observing the top 30 ranked Java projects on GitHub, they have, on average, 76% of classes with at least one annotation. Some projects may have more than 90% of its classes annotated. To measure code annotations usage and analyze their distribution, our work in [@LIMA2018] proposed a novel suite of software metrics dedicated to code annotations. To obtain threshold values, the Percentile Rank Analysis approach was used [@meirelles2013]. 
+Enterprise Java frameworks and APIs such as JPA (Java Persistence API), Spring, EJB (Enterprise Java Bean), and JUnit make extensive use of code annotations as means to allow applications to configure custom metadata and execute specific behavior. Observing the top 30 ranked Java projects on GitHub, they have, on average, 76% of classes with at least one annotation. Some projects may have more than 90% of its classes annotated. To measure code annotations usage and analyze their distribution, our work in [@LIMA2018] proposed a novel suite of software metrics dedicated to code annotations. We used a Percentile Rank Analysis approach [@meirelles2013] to obtain threshold values. 
 
-Source code metrics retrieve information from software to assess its characteristics. Well-known techniques use metrics associated with rules to detect bad smells on the source code [@Lanza2006]. However, traditional code metrics does not recognize code annotations on programming elements, which can lead to an incomplete code assessment [@Guerra2009]. For instance, a domain class can be considered simple using current complexity metrics. However, it can contain complex annotations for object-XML mapping. Also, using a set of annotations couples the application to a framework that can interpret them and current coupling metrics does not explicitly handle this.
+Source code metrics retrieve information from software to assess its characteristics. Well-known techniques use metrics associated with rules to detect bad smells on the source code [@Lanza2006]. However, traditional code metrics do not recognize code annotations on programming elements, which can lead to an incomplete code assessment [@Guerra2009]. For instance, a domain class can be considered simple using current complexity metrics. However, it can contain complex annotations for object-XML mapping. Also, using a set of annotations couples the application to a framework that can interpret them and current coupling metrics does not explicitly handle this.
 
-To automate the process of extracting the novel suite of software metrics for code annotation in [@LIMA2018], we developed an open source tool called Annotation Sniffer (`ASniffer`). It is a command line tool that reads java source code, extract the metrics values and ouputs an XML report. Potential `ASniffer` users are software engineers or researchers interested in static code analysis and mining software repositories. Additionally, given that it is an extensible tool, other developers can implement their own metrics and integrate them in the extraction process. Figure 1 presents a simple diagram of the ASniffer tool.
+To automate the process of extracting the novel suite of software metrics for code annotation in [@LIMA2018], we developed an open-source tool called Annotation Sniffer (`ASniffer`). It is a command-line tool that reads java source code, extracts the metrics values, and outputs an XML report. Potential `ASniffer` users are software engineers or researchers interested in static code analysis and mining software repositories. Additionally, given that it is an extensible tool, other developers can implement their metrics and integrate them in the extraction process. Figure 1 presents an overview diagram of the ASniffer tool.
 
 ![ASniffer Simple Diagram](figures/asniffer.png)
-Figure 1: ASniffer Simple Diagram
+Figure 1: ASniffer overview diagram
 
[…]
