### 2019-12-11 6ec99cb
Schema visible on the XML
Schema visible on the XML  for each annotation
