+++ src/main/java/br/inpe/cap/asniffer/metric/ASC.java
@@ -1,6 +1,7 @@
 package br.inpe.cap.asniffer.metric;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -22,7 +23,8 @@ import br.inpe.cap.asniffer.model.MetricResult;
 public class ASC extends ASTVisitor implements IClassMetricCollector {
 
 	List<String> imports = new ArrayList<>();
-	Set<String> schemas = new HashSet<>();
+	HashMap<String, String> schemasMapper = new HashMap<>();
+	CompilationUnit cu;
 	
 	@Override
 	public boolean visit(MarkerAnnotation node) {
@@ -45,28 +47,32 @@ public class ASC extends ASTVisitor implements IClassMetricCollector {
 	@Override
 	public void execute(CompilationUnit cu, MetricResult result, AMReport report) {
 		findImports(cu);
+		this.cu = cu;
 		cu.accept(this);
-		
-		
 	}
 
 	@Override
 	public void setResult(MetricResult result) {
 
-		result.setSchemas(schemas);
-		result.addClassMetric("ASC", schemas.size());
+		result.setSchemas(schemasMapper);
+		result.addClassMetric("ASC", result.getAnnotationSchemas().size());
 		
 	}
 	
 	private void findSchema(Annotation annotation) {
+		
 		for (String import_ : imports) {
 			if(import_.contains(annotation.getTypeName().getFullyQualifiedName())) {
 				int lastIndex = import_.lastIndexOf(".");
-				schemas.add(import_.substring(0,lastIndex));
+				schemasMapper.put(annotation.getTypeName().getFullyQualifiedName() + "-" +
+							cu.getLineNumber(annotation.getStartPosition())
+							,import_.substring(0,lastIndex));
+				return;
 			}
-			else
-				schemas.add("java.lang");
 		}
+		schemasMapper.put(annotation.getTypeName().getFullyQualifiedName() + "-" +
+				cu.getLineNumber(annotation.getStartPosition())
+	    ,"java.lang");
 	}
 	
 	private void findImports(CompilationUnit cu) {
