### 2019-12-17 #1923 @phillima
Thank you @gkthiruvathukal for accepting the role of editor of this paper, and thank you @arcuri82 for accepting the reviewer role. @gkthiruvathukal , I checked the list again and ran into these name for potential reviewers: @chrisma @danieledipompeo @sambitdash @fhg @mado89 .
