### 2019-12-17 #1923 @danieledipompeo
> Hi @danieledipompeo, you meant _review_, right? It is the holidays, so we understand you may need time. I'd like to aim for reviews to be complete sometime in January.

Yes, I can be a reviewer :). The deadline is fine with me.
