### 2019-12-17 #1923 @chrisma
Hi @gkthiruvathukal , thanks for considering me as a reviewer for this paper. This would be my first time reviewing for JOSS, so maybe being an additional, third reviewer would be a good start. Is this possible?
Thanks!
