### 2019-12-26 #1960 @arcuri82
hi,
 here some feedback, in no specific order of importance:

 - when I started the review, the repository had 0 stars.
   I starred it. You might want to do the same, as more stars give it a better impression.

 - "Download the source code and generate an executable jar file."
   That is not such a great approach.
   On GitHub, on the "release" page
   https://github.com/phillima/asniffer/releases
   you can add any custom file to a release, besides the zip of the source code.
   You could add `asniffer.jar` directly there as part of the release.

 - You might want to use Travis (https://travis-ci.org) or CircleCI (https://circleci.com).
   Using (the free) Travis is as simple as adding a `.travis.yml` file with content

   ```
    language: java
	sudo: false

	jdk:
	  - openjdk8

	cache:
	  directories:
		  - .autoconf
		  - $HOME/.m2

	notifications:
	  email: true

	script: mvn clean verify --fae
   ```

   and create an account then linked with a hook on GitHub (literally, it just take 5-10 minutes to do).
   You can then add a badge for Travis on your README.md file.
   For example, look at the GSON library from Google (https://github.com/google/gson),
   where the "build passing" is coming from Travis.
   Personally, I would not consider any open-source tool that does not use any form
   of Continuos Integration.

 - versioning of the project is inconsistent. The `pom.xml` is using version `0.0.1-SNAPSHOT`, whereas the release page has `2.0.0` as latest.

 - when building with `mvn package`, there is no generated file called `asniffer.jar`.
   There is instead: `asniffer-0.0.1-SNAPSHOT-jar-with-dependencies.jar`.
   You might want to change the configurations of the `maven-assembly-plugin` by adding `<finalName>asniffer</finalName>`

 - I tried to run the tool on itself:
   `java -jar target/asniffer-0.0.1-SNAPSHOT-jar-with-dependencies.jar  . .  single`
   The generated file is `..xml`. Is that really a good name?
   If I rather try with
   `java -jar target/asniffer-0.0.1-SNAPSHOT-jar-with-dependencies.jar  . foo.xml  single`
   I get a 	NullPointerException.
   Might want to provide bit more documentation on what `<path to xml report>` actually is.

 - `The path must be a root folder that contains other projects directories`. I guess a strong assumption here is that the projects are using Maven with `pom.xml` files, where source code is under `src/main/java`. This might not be true for other build tools such as Gradle and Ant. You might want to specify that you are considering projects structured with Maven conventions.

- ```
	mvn install
	mvn clean package
  ```
  bit confusing. Why doing an `install` before a `package`? You are creating an executable jar file, so package would suffice. Furthermore, even if you need `install`, that would execute `package` as well (see Maven lifecycle https://maven.apache.org/ref/3.6.3/maven-core/lifecycles.html). So, just a `mvn clean install` would suffice. Or am I misunderstanding something here?

- authorship of this publication needs some clarification. The authors are listed as:
  "Phyllipe Lima, Eduardo Guerra, and Paulo Meirelles".
  From the list of code contributors at
  https://github.com/phillima/asniffer/graphs/contributors
  There are 58 commits from @phillima (Phyllipe Lima?) and a single one from @paulormm (Paulo Meirelles?). It seems there are zero commits from Eduardo Guerra.
  Furthermore, the code is from a JSS paper titled `A Metrics Suite for code annotation assessment` (https://www.sciencedirect.com/science/article/pii/S016412121730273X), which has 6 authors, not 3.
  Please review the authorship guidelines at https://joss.readthedocs.io/en/latest/submitting.html#authorship
  which states: "Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are."
  This is the first time I review for JOSS, so maybe I am misunderstanding those JOSS rules. Could you please comment on them?
