### 2019-12-27 #1960 @phillima
Hello @arcuri82 , thank you very much for your comments and considerations. I will begin working on them.

- Regarding the "authorship", I would like to point out that @guerraem (Eduardo Guerra) and @paulormm (Paulo Meirelles) are my PhD Advisors and they were major contributors to the design and modelling of the tool. And also the definition and proposal of the metrics suite, the ones the ASniffer is able to extract.
The JSS paper (https://www.sciencedirect.com/science/article/pii/S016412121730273X) has other contributions (such as the threshold values and percentile rank analysis), which also involved other authors. However, not of all them were involved in the development and design of the ASniffer tool.
