### 2020-01-04 539c5e3
Adding CI with Travis
Adding CI with Travis
