### 2020-01-04 217c925
Path to XML Optional
The path to xml is optional. If no path is provided the ASniffer assumes the xml will be placed on the root folder of the project being analyzed.
