### 2020-01-04 227c490
Update .travis.yml
