+++ .travis.yml
@@ -1,12 +1,1 @@
  language: java
- sudo: false
-
- jdk:
-   - openjdk8
-
- cache:
-   directories:
- 	  - .autoconf
- 	  - $HOME/.m2
-
- script: mvn clean verify --fae
