+++ .travis.yml
@@ -1,1 +1,3 @@
  language: java
+ jdk:
+  - openjdk8
