+++ .travis.yml
@@ -1,3 +1,2 @@
  language: java
- jdk:
-  - openjdk8
+ jdk: openjdk8
