+++ .travis.yml
@@ -1,2 +1,12 @@
- language: java
- jdk: openjdk8
+language: java
+sudo: false
+
+jdk:
+  - openjdk8
+
+cache:
+ directories:
+  - .autoconf
+  - $HOME/.m2
+  
+ script: mvn clean verify --fae
