+++ README.md
@@ -7,11 +7,10 @@ Annotation Sniffer is a tool that extracts code annotation metrics from java sou
 
 ### How to download
 
-Download the source code and generate an executable jar file.
+Download the source code and generate an executable jar file. Or download the jar file provided with the latest release.
 
 ```
-mvn install
-mvn clean package
+mvn clean install
 ```
 ### How to use
 
@@ -19,8 +18,14 @@ mvn clean package
 java -jar asniffer.jar <path to project> <path to xml report> <single/multi>
 ```
 
-The path must be a root folder that contains other projects directories, for the "multi" case. 
-Follow the directory arrangement below.
+The "path to project" is mandatory, and should be the path to the java project to be analyzed (i.e, contains the source code files). Considering that only one java project is being analyzed, the directory should have the arrangement below.
+
+    .
+    ├── project                # Directory containing the source file for the project. This is the path provided
+
+In this case the ASniffer will consider that every ```.java``` file inside the directory ```project``` belongs to the same project.
+
+The ASniffer can also analyze multiple projects at once. In this case, the user should provide a directory with the arrangement described below.
 
     .
     ├── projects                # Root directory for projects. This is the path to be provided
@@ -28,13 +33,12 @@ Follow the directory arrangement below.
         ├── project2            # Contains the source files for project2
         └── ...         
 
-If the option "single" is used, then the path provided references only a single project, with all the source files inside that folder. 
+In this case the directory ```projects``` is a root folder, and the sub-directories ```project1```, ```project2``` and so forth, are each different java projects. They can be completely different projects. The user should manually arrange their projects directories to fit the arrangement described above, in order to use this ASniffer feature.
 
-    .
-    ├── project                # Directory containing the source file for the project. This is the path provided
-      
+The second parameter, "path to xml", is optional. It tells the ASniffer where to store the XML report file. If no path is provided, the ASniffer will place the report in the "path to project". This parameter is a path to a directoy and should not include any ```.xml``` in its name. The XML file will be generated by the ASniffer, with the projects name being the name of the XML file, i.e., ```projectsName.xml```. The ASniffer assumes that the name of the root directory is the name of the project. In the case several projects are being analyzed, the ASniffer considers that each sub-directory (inside the provided root directory) is the name of a separate project, and each project will have its own XML report placed in the "path to xml" (if provided, or in the "path to project" otherwise).
 
-For each project an xml output is generated with the report. The reports will be placed on the provided "path to xml".
+The third parameter (single/multi) informs the ASniffer if the "path to project" contains only one project (i.e, every ```.java``` file belongs to only one project) or several projects (i.e, the root directory contains several sub-directories, with each being a separate project).
+
 
 Annotation Metrics
 ==================
