### 2020-01-04 #1960 @phillima
Hello @arcuri82 . Once again, thank you for your comments and considerations. I have taken all off them in consideration and generated a new version of the tool (2.1.0). Please, feel free to make new comments and suggestions. Following is my response to your comments.

1. when I started the review, the repository had 0 stars.
I starred it. You might want to do the same, as more stars give it a better impression.
- **Response**: I have starred the project now. I have also asked @guerraem and @paulormm to do the same.

2. "Download the source code and generate an executable jar file."
That is not such a great approach.
On GitHub, on the "release" page
https://github.com/phillima/asniffer/releases
you can add any custom file to a release, besides the zip of the source code.
You could add asniffer.jar directly there as part of the release.
- **Response**: I have released a new version 2.1.0 (https://github.com/phillima/asniffer/releases) and included the executable jar file. I will keep on including the executable jar file as new versions are being released.

3. You might want to use Travis (https://travis-ci.org) or CircleCI (https://circleci.com).
- **Response**: I have integrated Travis CI with the ASniffer tool and added the badge in the Readme. Adding a form of CI definitely  improves the tool, and becomes much more professional.

4. Versioning of the project is inconsistent. The pom.xml is using version 0.0.1-SNAPSHOT, whereas the release page has 2.0.0 as latest.
-**Response**: I have updated the pom.xml to version 2.1.0 (the latest release).

5. When building with ```mvn package```, there is no generated file called ```asniffer.jar```.
-**Response**: I made the necessary adjustments in the pom.xml to fix this issue. The file is now correctly being generated as ```asniffer.jar```.

6. I tried to run the tool on itself: ```java -jar target/asniffer-0.0.1-SNAPSHOT-jar-with-dependencies.jar . . single```
The generated file is ```..xml```. Is that really a good name?
If I rather try with ```java -jar target/asniffer-0.0.1-SNAPSHOT-jar-with-dependencies.jar . foo.xml single``` I get a ```NullPointerException```.
Might want to provide bit more documentation on what <path to xml report> actually is.
-**Response**: Regarding the ```.xml``` file name, I fixed the part that generates the XML report. It was having trouble working around the ```.``` command as the ```current directory```. Using the ```normalize()``` method fixed it.

Regarding the ```foo.xml``` it will not work, because this is not how the ASniffer was intended to work. The "path to xml" should not include any ```.xml``` file name. It should only contain a path to a directory where you wish to store the report. Generating the XML file will be performed by the ASniffer. It will generate an xml file with the project's name. I changed the README file to make it clearer.
Also, the "path to xml" is now optional (as of version 2.1.0). If not included, the ASniffer will place the report in the "path to project".

7. - The path must be a root folder that contains other projects directories. I guess a strong assumption here is that the projects are using Maven with pom.xml files, where source code is under src/main/java. This might not be true for other build tools such as Gradle and Ant. You might want to specify that you are considering projects structured with Maven conventions.

-**Response**.: This multi-project feature assumes that the user wants to analyze several projects at once. It does not necessarily need to contain a ```pom.xml``` or be under ```src/main/java```.  However, if the user wants to use this feature is must manually arrange the folders as described in the README file. If the "single" option is used, the tool will search every .java file in the provided directory path, and consider it to be part of the same project. I have made updates in the README to make this clearer.

8. Why doing an install before a package?
-**Response**: I have updated the README file, and as you suggested ```mvn clean install``` is enough. It was indeed confusing.

Regarding the authorship I have explained in another comment.

Best Regards!
