### 2020-01-13 8e5ef49
Upgrade to log4j2
Upgrade to log4j2
