### 2020-01-27 #1960 @danieledipompeo
My comments follow.

### General checks

- **Contribution and authorship:** Has the submitting author (@phillima) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
It has been already clarify this point.

### Functionality

- **Installation:** Does installation proceed as outlined in the documentation?
I have installed the code on Ubuntu server 18.04, and a required plugin is missed:

```

<dependency>
    <groupId>javax.xml.bind</groupId>
    <artifactId>jaxb-api</artifactId>
    <version>2.4.0-b180830.0359</version>
</dependency>
```
- **Functionality:** Have the functional claims of the software been confirmed?

Yes, it has.

- **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

### Documentation

- **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

It is quite clear the point of this paper/tool. It extracts metrics from java code. Perhaps, a more in-depth description of competitors should be provided.

- **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The authors use maven to build the project. A needed dependency is missed. After adding it to the pom.xml, the project has been built.

- **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

No example is given.

- **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I cannot find any code documentation.

- **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

The source code is equipped with unit tests.

- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

### Software paper

- **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Yes, the paper provides itself enough information to understand the presented functionality.

- **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The presented tool extracts class metrics by exploiting the Eclipse JDT plugin. In the XML example file output (I suppose this file it has been generated through the tool) information about class metrics are listed.

- **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

No information is given about this statement. A more in-depth comparison should be provided.

- **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is short, but it is to follow.

- **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

It is quite extensive.
