### 2020-01-31 #1960 @phillima
Hello @danieledipompeo , thank you very much for your comments. I will work on them. I would like to point out some points.

- Regarding the missing maven dependency (JAXB), can you please confirm that you are using JDK 11 or higher? I checked that after JDK 11, some libraries (such as JAXB) were removed from the JDK. Since I developed the ASniffer on top of JDK 8 it was not necessary to add this dependency. To guarantee compatibility with future JDK versions, I will add this dependency on the pom file.

- Example Usage: In the README file there is the "How to use" section. I thought that would serve as "Example Usage". Can you confirm that this section should be improved to pass the "Example Usage"?

- State of the field: The reason we did not present a comparison with similar tools, is because the ASniffer collects Annotation Metrics. These metrics belong to a novel suite that we proposed on the paper [A Metrics Suite for Code Annoation Assessment](https://www.sciencedirect.com/science/article/pii/S016412121730273X). Hence there were no other tools to compare with.

As soon as I work on the other comments I will generate a new release.
