### 2020-02-01 #1960 @danieledipompeo
My comments below.

> Hello @danieledipompeo , thank you very much for your comments. I will work on them. I would like to point out some points.
>
> * Regarding the missing maven dependency (JAXB), can you please confirm that you are using JDK 11 or higher? I checked that after JDK 11, some libraries (such as JAXB) were removed from the JDK. Since I developed the ASniffer on top of JDK 8 it was not necessary to add this dependency. To guarantee compatibility with future JDK versions, I will add this dependency on the pom file.

Yes, if you use the JDK 11 you have to add the dependency, while if you use the JDK 8 you don't need because the JDK 8 is  "self-contained".

> * Example Usage: In the README file there is the "How to use" section. I thought that would serve as "Example Usage". Can you confirm that this section should be improved to pass the "Example Usage"?

Yes, if you add a running example or a simple one, it will be useful.

> * State of the field: The reason we did not present a comparison with similar tools, is because the ASniffer collects Annotation Metrics. These metrics belong to a novel suite that we proposed on the paper [A Metrics Suite for Code Annoation Assessment](https://www.sciencedirect.com/science/article/pii/S016412121730273X). Hence there were no other tools to compare with.

Ok, it's fine with me. It was just a suggestion. If you consider that your tool is the first one in this area you cannot find related work.
>
> As soon as I work on the other comments I will generate a new release.

Kind Regards
Daniele
