### 2020-02-01 9a7a196
Update README.md
Adding an Example Usage
