+++ README.md
@@ -206,18 +206,16 @@ The following example collects multiple projects, prints their name and the name
 
 ```java
 
-	List<AMReport> reports;
-	String path = "/projects";
+ List<AMReport> reports;
+ String path = "projects";
 
-	Runner runner = new Runner(path, path);
-	reports = runner.collectMultiple();
+ Runner runner = new Runner(path, path);
+ reports = runner.collectMultiple();
 
-	for (AMReport amReport : reports) {
-		System.out.println(amReport.getProjectName());
-
-		for (PackageModel packages : amReport.getPackages()) {
-			System.out.println(packages.getPackageName());
-		}
-	}
+ for (AMReport amReport : reports) {
+  System.out.println(amReport.getProjectName());
+  for (PackageModel packages : amReport.getPackages()) 
+  	System.out.println(packages.getPackageName());
+ }
 
 ```
