+++ src/main/java/br/inpe/cap/asniffer/ASniffer.java
@@ -5,6 +5,8 @@ import java.io.FileNotFoundException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import br.inpe.cap.asniffer.model.AMReport;
+
 
 public class ASniffer {
             
