+++ src/main/java/br/inpe/cap/asniffer/Runner.java
@@ -2,6 +2,8 @@ package br.inpe.cap.asniffer;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -25,22 +27,25 @@ public class Runner {
 	}
 	
 	//project path is a root directory to multiple project directories
-	public void collectMultiple(){
+	public List<AMReport> collectMultiple(){
+		List<AMReport> reports = new ArrayList<AMReport>();
 		for (Path projectPath : FileUtils.getProjectsPath(projectsPath)) 
-			collect(projectPath);
+			reports.add(collect(projectPath));
+		return reports;
 	}
 	
 	//project path is a directory to a single
-	public void collectSingle() {
-		collect(Paths.get(projectsPath));
+	public AMReport collectSingle() {
+		return collect(Paths.get(projectsPath));
 	}
 	
-	private void collect(Path projectPath) {
+	private AMReport collect(Path projectPath) {
 		String projectName = FileUtils.getProjectName(projectPath);
 		logger.info("Initializing extraction for project " + projectName);
 		AMReport report = new AM().calculate(projectPath.toString(), projectName);
 		logger.info("Extraction concluded for project " + projectName);
 		IReport xmlReport = new XMLReport();
 		xmlReport.generateReport(report, xmlPath);
+		return report;
 	}
 }
