+++ joss-paper/paper.bib
@@ -116,3 +116,19 @@ keywords = "Framework, Metadata, Metadata-based framework, Software architecture
     MONTH="sep",
     YEAR="2018"
 }
+
+@manual{aniche2015,
+  title={Java code metrics calculator (CK)},
+  author={Maurício Aniche},
+  year={2015},
+  note={Available in https://github.com/mauricioaniche/ck/}
+}
+
+@INPROCEEDINGS{ck91,
+  author = {Shyam R. Chidamber and Chris F. Kemerer},
+  title = {Towards a Metrics Suite for Object-Oriented Design},
+  booktitle = {Proceedings of OOPSLA'91},
+  year = {1991},
+  pages = {197--211},
+  ordinal = {P422}
+}
