+++ README.md
@@ -200,7 +200,7 @@ If you wish to use the ASniffer as an API on your own projects, we provide some 
 
 With these calls, the ASniffer will run, collect the annotation metrics, and place the XML report file on the provided path. However, if you would like to perform some analysis on the metrics values, both ```collectSingle()``` and ```collectMultiple()``` return, respectively, an instance of ```AMReport``` and a ```List<AMReport>```. The ```AMReport``` class contains the complete report of the collected metrics for each Java project. For this reason the ```collectMultiple()``` returns a list of ```AMReports``` (one for each Java project). 
 
-The ```AMReport``` located [here](https://github.com/phillima/asniffer/blob/master/src/main/java/br/inpe/cap/asniffer/model/AMReport.java) contains the project's name, and the list of packages. The packages are stored in the ```PackageModel``` class, which in turn contains the ```MetricResult``` class that stores the annotation metrics values.
+The ```AMReport``` (located [here](https://github.com/phillima/asniffer/blob/master/src/main/java/br/inpe/cap/asniffer/model/AMReport.java)) contains the project's name, and the list of packages. The packages are stored in the ```PackageModel``` class, which in turn contains the ```MetricResult``` class that stores the annotation metrics values.
 
 The following example collects multiple projects, prints their name and the name of every package.
 
