### 2020-02-01 #1960 @phillima
Hi @danieledipompeo, once again, thank you for your comments. I have generated a new version of the ASniffer tool (2.2.0) that aims to address your review. Please, feel free to make new comments and suggestions.

> ### Functionality
> * **Installation:** Does installation proceed as outlined in the documentation?
>   I have installed the code on Ubuntu server 18.04, and a required plugin is missed:
>
> ```
>
> <dependency>
>     <groupId>javax.xml.bind</groupId>
>     <artifactId>jaxb-api</artifactId>
>     <version>2.4.0-b180830.0359</version>
> </dependency>
> ```

- I have added this missing dependency to the pom.xml file, to make the ASniffer compatible with the Java version 11 and higher.

> * **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> The authors use maven to build the project. A needed dependency is missed. After adding it to the pom.xml, the project has been built.

- The missing dependency was added!

> * **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> No example is given.
>

- In the Readme file I have added a new section "Example Usage". In this section I show how to extract annotation metrics from a Java project using the own ASniffer code as sample. I have also uploaded the generated XML file to the repository to serve as an example.

> * **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> I cannot find any code documentation.
>

- To address this issue I have added in the Readme file a new section "ASniffer API". I explain and demonstrate how a developer can use the ASniffer as an API in his own project.

> ### Software paper
> * **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> No information is given about this statement. A more in-depth comparison should be provided.
>

- To address this comment I have added in the paper a new section called "Related Work". I explain that the ASniffer collects metrics from a novel suite, so no similar tool was found to compare. However, there are other tools that perform static analysis to collect source code metrics. I make a reference to an open source tool called [CK Tool](https://github.com/mauricioaniche/ck/), that also collect source code metrics and uses the JDT as the AST parser.

Kind Regards!
Phyllipe
