### 2020-02-17 #1960 @phillima
Hi @arcuri82. I am glad I have addressed your concerns! Thank you
Hi @danieledipompeo. After your first review I generated a new version of the tool (2.2.0) that hopefully addressed your concerns. Please, feel free to make new comments.
