### 2020-03-08 #1960 @gkthiruvathukal
@danieledipompeo Just checking whether your issues have been addressed by @phillima. I think almost all of your and @arcuri82  issues have been addressed.  I'd like to move toward acceptance.
