+++ README.md
@@ -1,5 +1,5 @@
 [![Build Status](https://travis-ci.com/phillima/asniffer.svg?branch=master)](https://travis-ci.com/phillima/asniffer)
-
+[![DOI](https://zenodo.org/badge/76030375.svg)](https://zenodo.org/badge/latestdoi/76030375)
 Annotation Sniffer
 ==================
 
