### 2020-03-10 #1960 @phillima
Hello @gkthiruvathukal. I have performed the steps as requested. Could you please verify that I have done it correctly? Thank you very much!

> Ok, I am ready to proceed toward acceptance. @phillima here is a final checklist for you.
>
> Please do the following:
>
> * [X ]  Make a tagged release of your software, and list the version tag of the archived version here.
Version tag: 2.2.1
> * [ X]  Archive the reviewed software in Zenodo
> * [ X]  Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
> * [ X]  List the Zenodo DOI of the archived version here.
Zenodo DOI: 10.5281/zenodo.3701910
